requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.27.0
sentence-transformers>=2.2.0
easyocr>=1.7.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
scikit-learn>=1.3.0
tqdm>=4.65.0
Pillow>=9.5.0
numpy>=1.24.0
matplotlib>=3.7.0